refactor: split linter's functions into multiple files
